Expose enumeration of I2c funcs
Previously, smbus.funcs was defined, but it's flags were not exported.

To avoid clutter, I've moved all the values into the `I2cFunc` class, and exported it.

On newer python releases, I2cFunc acts as an IntFlags, which is easier to read than a plain int.
On older releases, it falls back to a pure `int`.